+++ omod/src/main/java/org/openmrs/module/appointments/web/controller/RecurringAppointmentsController.java
@@ -2,6 +2,7 @@ package org.openmrs.module.appointments.web.controller;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.openmrs.api.APIException;
 import org.openmrs.module.appointments.model.Appointment;
 import org.openmrs.module.appointments.model.AppointmentRecurringPattern;
 import org.openmrs.module.appointments.service.AppointmentRecurringPatternService;
@@ -10,6 +11,7 @@ import org.openmrs.module.appointments.web.contract.RecurringPattern;
 import org.openmrs.module.appointments.web.helper.RecurringPatternHelper;
 import org.openmrs.module.appointments.web.mapper.AbstractAppointmentRecurringPatternMapper;
 import org.openmrs.module.appointments.web.mapper.AppointmentMapper;
+import org.openmrs.module.appointments.web.validators.impl.RecurringPatternValidator;
 import org.openmrs.module.webservices.rest.web.RestConstants;
 import org.openmrs.module.webservices.rest.web.RestUtil;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +19,8 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.validation.BeanPropertyBindingResult;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -25,6 +29,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Objects;
 
 @Controller
 @RequestMapping(value = "/rest/" + RestConstants.VERSION_1 + "/recurring-appointments")
@@ -34,6 +39,9 @@ public class RecurringAppointmentsController {
 
     @Autowired
     private AppointmentRecurringPatternService appointmentRecurringPatternService;
+
+    @Autowired
+    private RecurringPatternValidator recurringPatternValidator;
 
     @Autowired
     private RecurringPatternHelper recurringPatternHelper;
@@ -50,7 +58,11 @@ public class RecurringAppointmentsController {
     public ResponseEntity<Object> save(@RequestBody AppointmentRequest appointmentRequest) {
         try {
             RecurringPattern recurringPattern = appointmentRequest.getRecurringPattern();
-            recurringPatternHelper.validateRecurringPattern(recurringPattern);
+            Errors errors = new BeanPropertyBindingResult(recurringPattern, "recurringPattern");
+            recurringPatternValidator.validate(recurringPattern, errors);
+            if (!Objects.isNull(errors) && !errors.getAllErrors().isEmpty()) {
+                throw new APIException(errors.getAllErrors().get(0).getCodes()[1]);
+            }
             AppointmentRecurringPattern appointmentRecurringPattern = allAppointmentRecurringPatternMapper.fromRequest(recurringPattern);
             List<Appointment> appointmentsList = recurringPatternHelper.generateRecurringAppointments(appointmentRecurringPattern,
                     appointmentRequest);
@@ -63,6 +75,4 @@ public class RecurringAppointmentsController {
             return new ResponseEntity<>(RestUtil.wrapErrorResponse(e, e.getMessage()), HttpStatus.BAD_REQUEST);
         }
     }
-
-
 }
