+++ omod/src/test/java/org/openmrs/module/appointments/web/controller/RecurringAppointmentsControllerTest.java
@@ -13,7 +13,11 @@ import org.openmrs.module.appointments.web.contract.RecurringPattern;
 import org.openmrs.module.appointments.web.helper.RecurringPatternHelper;
 import org.openmrs.module.appointments.web.mapper.AbstractAppointmentRecurringPatternMapper;
 import org.openmrs.module.appointments.web.mapper.AppointmentMapper;
+import org.openmrs.module.appointments.web.validators.impl.RecurringPatternValidator;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.Errors;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -22,6 +26,9 @@ import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.MockitoAnnotations.initMocks;
 import static org.powermock.api.mockito.PowerMockito.when;
@@ -36,6 +43,9 @@ public class RecurringAppointmentsControllerTest {
 
     @Mock
     private RecurringPatternHelper recurringPatternHelper;
+
+    @Mock
+    private RecurringPatternValidator recurringPatternValidator;
 
     @Mock
     @Qualifier("allAppointmentRecurringPatternMapper")
@@ -53,9 +63,6 @@ public class RecurringAppointmentsControllerTest {
     public void shouldSaveAnAppointmentWhenRecurringPatternAlsoPresentsInRequest() {
         AppointmentRequest appointmentRequest = new AppointmentRequest();
         RecurringPattern recurringPattern = new RecurringPattern();
-        recurringPattern.setType("sometype");
-        recurringPattern.setPeriod(1);
-        recurringPattern.setFrequency(2);
         appointmentRequest.setStartDateTime(new Date());
         appointmentRequest.setUuid("someUuid");
         appointmentRequest.setServiceUuid("someServiceUuid");
@@ -71,13 +78,17 @@ public class RecurringAppointmentsControllerTest {
         when(appointmentRecurringPatternService.validateAndSave(appointmentRecurringPattern)).thenReturn(appointments);
         when(recurringPatternHelper.generateRecurringAppointments(appointmentRecurringPattern, appointmentRequest)).thenReturn(new ArrayList<>());
         when(appointmentMapper.constructResponse(Arrays.asList(appointmentOne, appointmentTwo))).thenReturn(Collections.emptyList());
+        doNothing().when(recurringPatternValidator).validate(recurringPattern, mock(Errors.class));
         appointmentRecurringPattern.setAppointments(Collections.emptySet());
 
-        recurringAppointmentsController.save(appointmentRequest);
+        ResponseEntity<Object> responseEntity = recurringAppointmentsController.save(appointmentRequest);
 
         Mockito.verify(allAppointmentRecurringPatternMapper, times(1)).fromRequest(recurringPattern);
         Mockito.verify(appointmentRecurringPatternService, times(1)).validateAndSave(appointmentRecurringPattern);
         Mockito.verify(recurringPatternHelper, times(1)).generateRecurringAppointments(appointmentRecurringPattern, appointmentRequest);
         Mockito.verify(appointmentMapper, times(1)).constructResponse(Collections.emptyList());
+        assertEquals(responseEntity.getStatusCode(), HttpStatus.OK);
     }
+
+    // TODO : Write a test when there are errors
 }
