+++ omod/src/test/java/org/openmrs/module/appointments/web/helper/RecurringPatternHelperTest.java
@@ -38,153 +38,6 @@ public class RecurringPatternHelperTest {
         recurringPatternHelper = new RecurringPatternHelper(appointmentMapperMock);
     }
 
-    private RecurringPattern getRecurringPattern() {
-        RecurringPattern recurringPattern = new RecurringPattern();
-        recurringPattern.setType("DAY");
-        recurringPattern.setPeriod(1);
-        recurringPattern.setFrequency(1);
-        return recurringPattern;
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenTypeIsNullInRecurringPattern() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setType(null);
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenTypeIsEmptyInRecurringPattern() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setType("");
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenInvalidType() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setType("TEXT");
-        recurringPattern.setDaysOfWeek(null);
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenPeriodIsLessThanOneInRecurringPattern() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setPeriod(0);
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenFrequencyIsLessThanOneInRecurringPattern() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setFrequency(0);
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenFrequencyAndEndDateAreNullInRecurringPattern() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setFrequency(null);
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    private String getExceptionMessage() {
-        return "type should be DAY/WEEK\n" +
-                "period and frequency/endDate are mandatory if type is DAY\n" +
-                "daysOfWeek, period and frequency/endDate are mandatory if type is WEEK";
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenEndDateIsNull() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setEndDate(null);
-        recurringPattern.setFrequency(0);
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenWeekTypeWithEmptyDays() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setType("WEEk");
-        recurringPattern.setEndDate(null);
-        recurringPattern.setFrequency(2);
-        recurringPattern.setPeriod(2);
-        recurringPattern.setDaysOfWeek(Arrays.asList());
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenWeekTypeWithNoDays() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setType("WEEk");
-        recurringPattern.setEndDate(null);
-        recurringPattern.setFrequency(2);
-        recurringPattern.setPeriod(2);
-        recurringPattern.setDaysOfWeek(null);
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenWeekTypeWithInvalidDays() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setType("WEEk");
-        recurringPattern.setFrequency(2);
-        recurringPattern.setPeriod(2);
-        recurringPattern.setDaysOfWeek(Arrays.asList("TEXT", "MISC"));
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenWeekTypeWithNoPeriod() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setType("WEEk");
-        recurringPattern.setFrequency(2);
-        recurringPattern.setPeriod(0);
-        recurringPattern.setDaysOfWeek(Arrays.asList("SUNDAY", "MONDAY"));
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenWeekTypeWithNoFrequency() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        recurringPattern.setType("WEEk");
-        recurringPattern.setFrequency(0);
-        recurringPattern.setPeriod(2);
-        recurringPattern.setDaysOfWeek(Arrays.asList("SUNDAY", "MONDAY"));
-        expectedException.expect(APIException.class);
-        expectedException.expectMessage(getExceptionMessage());
-        recurringPatternHelper.validateRecurringPattern(recurringPattern);
-    }
-
-    @Test
-    public void shouldNotThrowAnExceptionForValidData() {
-        RecurringPattern recurringPattern = getRecurringPattern();
-        assertTrue(recurringPatternHelper.validateRecurringPattern(recurringPattern));
-    }
-
     private AppointmentRequest getAppointmentRequest(Date appointmentStartDateTime, Date appointmentEndDateTime) {
         AppointmentRequest appointmentRequest = new AppointmentRequest();
         appointmentRequest.setStartDateTime(appointmentStartDateTime);
